# apps/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Flowza API",
    "version": "0.1.0",
    "description": "Документация по уже реализованным backend-эндпоинтам Flowza. Спецификация собрана по текущим NestJS controller и DTO."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local backend"
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Зарегистрировать пользователя",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" },
              "example": {
                "phone": "+79991234567",
                "login": "tokyo_roll_user",
                "email": "user@example.com",
                "password": "StrongPass123",
                "consentToPrivacyPolicy": true,
                "consentToPersonalData": true,
                "agreementVersion": "2026-04",
                "captchaToken": "turnstile-token"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Регистрация успешна, требуется подтверждение email",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegisterResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ErrorResponse" },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "409": { "$ref": "#/components/responses/ErrorResponse" },
          "429": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Войти по телефону, email или login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" },
              "example": {
                "identifier": "admin@example.com",
                "password": "StrongPass123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Успешный вход",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "429": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/auth/register/verify-email": {
      "post": {
        "tags": ["Auth"],
        "summary": "Подтвердить email кодом из письма",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyEmailCodeDto" },
              "example": {
                "email": "user@example.com",
                "code": "123456"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Email подтвержден",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VerifyEmailResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "429": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/auth/register/resend-email-code": {
      "post": {
        "tags": ["Auth"],
        "summary": "Запросить повторную отправку email-кода",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResendEmailCodeDto" },
              "example": {
                "email": "user@example.com"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Повторная отправка обработана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResendEmailCodeResponse"
                }
              }
            }
          },
          "429": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Получить текущего пользователя",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Данные текущего пользователя",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorResponse" }
        }
      },
      "delete": {
        "tags": ["Auth"],
        "summary": "Деактивировать текущего пользователя",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Пользователь деактивирован",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GenericSuccess" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Получить товары организации",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Товары организации",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationProduct"
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "403": { "$ref": "#/components/responses/ErrorResponse" }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Создать товар",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProductDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Товар создан",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrganizationProduct" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "403": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/products/{productId}": {
      "patch": {
        "tags": ["Products"],
        "summary": "Обновить товар",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "example": 7
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProductDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Товар обновлен",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrganizationProduct" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "403": { "$ref": "#/components/responses/ErrorResponse" },
          "404": { "$ref": "#/components/responses/ErrorResponse" }
        }
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Удалить товар",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "example": 7
          }
        ],
        "responses": {
          "200": {
            "description": "Товар удален",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer", "example": 7 }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "403": { "$ref": "#/components/responses/ErrorResponse" },
          "404": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Ошибка запроса",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "statusCode": { "type": "integer", "example": 400 },
          "message": {
            "type": "string",
            "example": "Ошибка валидации"
          },
          "errorCode": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string", "example": "email" },
                "message": {
                  "type": "string",
                  "example": "Email должен быть валидным адресом"
                }
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-29T18:00:00.000Z"
          },
          "path": { "type": "string", "example": "/auth/register" }
        },
        "required": ["success", "statusCode", "message", "errorCode", "timestamp", "path"]
      },
      "GenericSuccess": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true }
        },
        "required": ["success"]
      },
      "UserRole": {
        "type": "string",
        "enum": ["superAdmin", "admin", "moderator", "operator", "user"]
      },
      "RegisterDto": {
        "type": "object",
        "required": [
          "phone",
          "login",
          "email",
          "password",
          "consentToPrivacyPolicy",
          "consentToPersonalData",
          "agreementVersion",
          "captchaToken"
        ],
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^\\+7\\d{10}$",
            "example": "+79991234567"
          },
          "login": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]+$",
            "example": "tokyo_roll_user"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "StrongPass123"
          },
          "consentToPrivacyPolicy": { "type": "boolean", "example": true },
          "consentToPersonalData": { "type": "boolean", "example": true },
          "agreementVersion": { "type": "string", "example": "2026-04" },
          "captchaToken": { "type": "string", "example": "turnstile-token" }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "required": [
          "success",
          "message",
          "emailSentViaSmtp",
          "verificationRequired",
          "verificationTtlSec",
          "resendAvailableInSec"
        ],
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": {
            "type": "string",
            "example": "Код подтверждения отправлен на email"
          },
          "emailSentViaSmtp": { "type": "boolean", "example": true },
          "verificationRequired": { "type": "boolean", "example": true },
          "verificationTtlSec": { "type": "integer", "example": 900 },
          "resendAvailableInSec": { "type": "integer", "example": 60 }
        }
      },
      "LoginDto": {
        "type": "object",
        "required": ["identifier", "password"],
        "properties": {
          "identifier": { "type": "string", "example": "admin@example.com" },
          "password": { "type": "string", "example": "StrongPass123" }
        }
      },
      "AuthUser": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true,
            "example": "user@example.com"
          },
          "phone": { "type": "string", "example": "+79991234567" },
          "login": { "type": "string", "example": "tokyo_roll_user" },
          "role": { "$ref": "#/components/schemas/UserRole" },
          "primaryTenantId": {
            "type": "integer",
            "nullable": true,
            "example": 1
          },
          "organizationIds": {
            "type": "array",
            "items": { "type": "integer" },
            "example": [1, 2]
          },
          "isActive": { "type": "boolean", "example": true }
        },
        "required": [
          "id",
          "phone",
          "login",
          "role",
          "primaryTenantId",
          "organizationIds"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "required": ["accessToken", "user"],
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
          },
          "user": { "$ref": "#/components/schemas/AuthUser" }
        }
      },
      "VerifyEmailCodeDto": {
        "type": "object",
        "required": ["email", "code"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "example": "123456"
          }
        }
      },
      "VerifyEmailResponse": {
        "type": "object",
        "required": ["success", "accessToken", "user"],
        "properties": {
          "success": { "type": "boolean", "example": true },
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
          },
          "user": { "$ref": "#/components/schemas/AuthUser" }
        }
      },
      "ResendEmailCodeDto": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          }
        }
      },
      "ResendEmailCodeResponse": {
        "type": "object",
        "required": ["success", "message", "resendAvailableInSec"],
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": {
            "type": "string",
            "example": "Если аккаунт существует, код был отправлен"
          },
          "emailSentViaSmtp": { "type": "boolean", "example": true },
          "resendAvailableInSec": { "type": "integer", "example": 60 }
        }
      },
      "MeResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/AuthUser" },
          {
            "type": "object",
            "required": ["isActive"],
            "properties": {
              "isActive": { "type": "boolean", "example": true }
            }
          }
        ]
      },
...
      "OrganizationProduct": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 7 },
          "tenantId": { "type": "integer", "example": 1 },
          "categoryId": { "type": "integer", "example": 5 },
          "name": { "type": "string", "example": "Филадельфия" },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Лосось, сыр, рис"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/product.jpg"
          },
          "badgeText": { "type": "string", "nullable": true, "example": "Хит" },
          "price": { "type": "integer", "example": 520 },
          "currency": { "type": "string", "example": "RUB" },
          "isActive": { "type": "boolean", "example": true }
        },
        "required": [
          "id",
          "tenantId",
          "categoryId",
          "name",
          "description",
          "imageUrl",
          "badgeText",
          "price",
          "currency",
          "isActive"
        ]
      },
...
      "CreateProductDto": {
        "type": "object",
        "required": ["tenantId", "categoryId", "name", "price"],
        "properties": {
          "tenantId": { "type": "integer", "example": 1 },
          "categoryId": { "type": "integer", "example": 5 },
          "name": { "type": "string", "example": "Филадельфия" },
          "description": { "type": "string", "example": "Лосось, сыр, рис" },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/product.jpg"
          },
          "badgeText": { "type": "string", "example": "Хит" },
          "price": { "type": "integer", "example": 520 },
          "currency": { "type": "string", "example": "RUB" },
          "isActive": { "type": "boolean", "example": true }
        }
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "categoryId": { "type": "integer", "example": 6 },
          "name": { "type": "string", "example": "Филадельфия люкс" },
          "description": { "type": "string", "example": "Больше лосося" },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/product-new.jpg"
          },
          "badgeText": { "type": "string", "example": "Новинка" },
          "price": { "type": "integer", "example": 590 },
          "currency": { "type": "string", "example": "RUB" },
          "isActive": { "type": "boolean", "example": true }
        }
      },
...
    }
  }
}